remove more integration tests, server tests type check
