+++ package.json
@@ -94,13 +94,12 @@
     "@types/isomorphic-fetch": "0.0.34",
     "@types/jasmine": "^2.8.2",
     "@types/jest": "21.1.8",
-    "@types/lodash": "4.14.88",
+    "@types/lodash": "4.14.89",
     "@types/object-assign": "4.0.30",
-    "@types/react": "16.0.29",
+    "@types/react": "16.0.30",
     "@types/react-addons-test-utils": "0.14.20",
     "@types/react-dom": "16.0.3",
-    "@types/react-redux": "5.0.14",
-    "@types/sinon": "4.1.1",
+    "@types/sinon": "4.1.2",
     "@types/zen-observable": "0.5.3",
     "apollo-cache-inmemory": "1.1.4",
     "apollo-client": "2.1.0",
@@ -125,8 +124,6 @@
     "lint-staged": "6.0.0",
     "lodash": "4.17.4",
     "minimist": "1.2.0",
-    "mobx": "3.4.1",
-    "mobx-react": "4.3.5",
     "pre-commit": "1.2.2",
     "prettier": "1.9.2",
     "pretty-bytes": "4.0.2",
@@ -134,10 +131,8 @@
     "react-addons-test-utils": "15.6.2",
     "react-dom": "16.2.0",
     "react-native": "0.51.0",
-    "react-redux": "5.0.6",
     "react-test-renderer": "16.2.0",
     "recompose": "0.26.0",
-    "redux": "3.7.2",
     "rimraf": "^2.6.2",
     "rollup": "0.52.1",
     "source-map-support": "0.5.0",
